Change some variables from float to double to avoid thousands of compiler warnings on Windows

+++ opticsApp/src/chantler.h
@@ -15,10 +15,10 @@ double density[NUM_SPECIES] = {
 struct filterMaterial {
 	int Z;
 	char *name;
-	float density;
+	double density;
 	int numEntries;
-	float keV[NUM_ENTRIES];
-	float mu[NUM_ENTRIES];
+	double keV[NUM_ENTRIES];
+	double mu[NUM_ENTRIES];
 } filt[NUM_SPECIES] = {
 	{
 		4,
